+++ application/controllers/Department.php
@@ -20,7 +20,17 @@ class Department extends CI_Controller {
 	 */
 	public function index()
 	{
-		$this->load->template('department/index','',1);
+		
+		$role_id = 1;
+		
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='department' and me_menu_access.view_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		if(!empty($data["access"])){
+			$this->load->template('department/index','',1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
 	}
 	public function response(){
 		$draw = intval($this->input->get("draw"));
@@ -28,15 +38,26 @@ class Department extends CI_Controller {
 		$length = intval($this->input->get("length"));
 		$data["department"] = $this->departmentmodal->ShowDepartment();
 		$data1 = [];
-		$count=0;     
+		$count=0;  
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='department' and me_menu_access.edit_access =1";
+        $data["edit_access"] = $this->menumodal->ShowMenuBySearch($condition);
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='department' and me_menu_access.delete_access =1";
+        $data["delete_access"] = $this->menumodal->ShowMenuBySearch($condition);
+		   
 		foreach($data["department"] as $r) {
 			$DeparmentID = $r["DeparmentID"];
-			$btn = '<button type="button" class="btn btn-soft-primary waves-effect waves-light">';
+			$btn = '';
+			if(!empty($data["edit_access"])){
+			$btn .= '<button type="button" class="btn btn-soft-primary waves-effect waves-light">';
 			$btn .= '<i class="mdi mdi-pencil font-size-16 align-middle" style="line-height: 1;"></i>';
 			$btn .= '</button>';
+			}
+			if(!empty($data["delete_access"])){
 			$btn .= '<button onClick="delete_department(this.id)" id="'.$DeparmentID.'" type="button" class="btn btn-soft-danger waves-effect waves-light">';
 			$btn .= '<i class="bx bx-block font-size-16 align-middle"></i>';
 			$btn .= '</button>';
+			}
 			$data1[] = array(
 				++$count,
 				$r["DepartmentName"],
@@ -53,7 +74,17 @@ class Department extends CI_Controller {
 		echo json_encode($result);
 	}
 	public function add(){
-		$this->load->template('department/add','',1);
+		$role_id = 1;
+		
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='department' and me_menu_access.add_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		if(!empty($data["access"])){
+			$this->load->template('department/add','',1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
+		
 	}
 	public function process(){
 		$data["save"] = $this->departmentmodal->AddDepartment($_POST);
