+++ application/controllers/Role.php
@@ -20,7 +20,17 @@ class Role extends CI_Controller {
 	 */
 	public function index()
 	{
-		$this->load->template('role/index','',1);
+		$role_id = 1;
+		
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='role' and me_menu_access.view_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		if(!empty($data["access"])){
+			$this->load->template('role/index','',1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
+		
 	}
 	public function response(){
 		$draw = intval($this->input->get("draw"));
@@ -28,15 +38,24 @@ class Role extends CI_Controller {
 		$length = intval($this->input->get("length"));
 		$data["role"] = $this->rolemodal->ShowRole();
 		$data1 = [];
-		$count=0;     
+		$count=0;    
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='role' and me_menu_access.edit_access =1";
+        $data["edit_access"] = $this->menumodal->ShowMenuBySearch($condition);
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='role' and me_menu_access.delete_access =1";
+        $data["delete_access"] = $this->menumodal->ShowMenuBySearch($condition);
 		foreach($data["role"] as $r) {
 			$RoleID = $r["RoleID"];
+			if(!empty($data["edit_access"])){
 			$btn = '<button type="button" class="btn btn-soft-primary waves-effect waves-light">';
 			$btn .= '<i class="mdi mdi-pencil font-size-16 align-middle" style="line-height: 1;"></i>';
 			$btn .= '</button>';
+			}
+			if(!empty($data["delete_access"])){
 			$btn .= '<button onClick="delete_role(this.id)" id="'.$RoleID.'" type="button" class="btn btn-soft-danger waves-effect waves-light">';
 			$btn .= '<i class="bx bx-block font-size-16 align-middle"></i>';
 			$btn .= '</button>';
+			}
 			$data1[] = array(
 				++$count,
 				$r["DepartmentName"],
@@ -60,8 +79,17 @@ class Role extends CI_Controller {
 		}
 	}
 	public function add(){
-		$data["department"] = $this->departmentmodal->ShowDepartment();
-		$this->load->template('role/add',$data,1);
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='role' and me_menu_access.add_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		if(!empty($data["access"])){
+			$data["department"] = $this->departmentmodal->ShowDepartment();
+			$this->load->template('role/add',$data,1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
+		
 	}
 	public function process(){
 		$data["save"] = $this->rolemodal->AddRole($_POST);
